Commit de atualizacao de algumas funcoes integrando-as com o mysql
